Fix, Refactor(DK-515): 프로필 클릭 시 열리는 모달 아이템 클릭시 모달 닫히도록 수정 및 컴포넌트 분리
- HeaderMenuListItem 컴포넌트 분리
